+++ src/backoffice/dashboards/published-status/dashboard-published-status.test.ts
@@ -1,6 +1,6 @@
 import { expect, fixture, html } from '@open-wc/testing';
 
-import { defaultA11yConfig } from '../../../core/helpers/chai';
+import { defaultA11yConfig } from '../../../core/test-utils/chai';
 import { UmbDashboardPublishedStatusElement } from './dashboard-published-status.element';
 
 describe('UmbDashboardPublishedStatus', () => {
