+++ src/backoffice/dashboards/telemetry/dashboard-telemetry.test.ts
@@ -1,5 +1,5 @@
 import { expect, fixture, html } from '@open-wc/testing';
-import { defaultA11yConfig } from '../../../core/helpers/chai';
+import { defaultA11yConfig } from '../../../core/test-utils/chai';
 import { UmbDashboardTelemetryElement } from './dashboard-telemetry.element';
 
 describe('UmbDashboardTelemetryElement', () => {
