+++ src/backoffice/property-editor-uis/block-list/property-editor-ui-block-list.test.ts
@@ -1,5 +1,5 @@
 import { expect, fixture, html } from '@open-wc/testing';
-import { defaultA11yConfig } from '../../../core/helpers/chai';
+import { defaultA11yConfig } from '../../../core/test-utils/chai';
 import { UmbPropertyEditorUIBlockListElement } from './property-editor-ui-block-list.element';
 
 describe('UmbPropertyEditorUIBlockListElement', () => {
