+++ src/backoffice/property-editor-uis/checkbox-list/property-editor-ui-checkbox-list.test.ts
@@ -1,5 +1,5 @@
 import { expect, fixture, html } from '@open-wc/testing';
-import { defaultA11yConfig } from '../../../core/helpers/chai';
+import { defaultA11yConfig } from '../../../core/test-utils/chai';
 import { UmbPropertyEditorUICheckboxListElement } from './property-editor-ui-checkbox-list.element';
 
 describe('UmbPropertyEditorUICheckboxListElement', () => {
