+++ src/backoffice/property-editor-uis/icon-picker/property-editor-ui-icon-picker.test.ts
@@ -1,5 +1,5 @@
 import { expect, fixture, html } from '@open-wc/testing';
-import { defaultA11yConfig } from '../../../core/helpers/chai';
+import { defaultA11yConfig } from '../../../core/test-utils/chai';
 import { UmbPropertyEditorUIIconPickerElement } from './property-editor-ui-icon-picker.element';
 
 describe('UmbPropertyEditorUIIconPickerElement', () => {
