+++ src/backoffice/property-editor-uis/toggle/property-editor-ui-toggle.test.ts
@@ -1,5 +1,5 @@
 import { expect, fixture, html } from '@open-wc/testing';
-import { defaultA11yConfig } from '../../../core/helpers/chai';
+import { defaultA11yConfig } from '../../../core/test-utils/chai';
 import { UmbPropertyEditorUIToggleElement } from './property-editor-ui-toggle.element';
 
 describe('UmbPropertyEditorUIToggleElement', () => {
