+++ src/core/mocks/domains/telemetry.handlers.ts
@@ -1,6 +1,6 @@
 import { rest } from 'msw';
 
-import umbracoPath from '../../helpers/umbraco-path';
+import umbracoPath from '../../utils/umbraco-path';
 import type { ConsentLevelSettings, TelemetryModel } from '@umbraco-cms/models';
 
 let telemetryLevel: TelemetryModel['level'] = 'Basic';
