+++ src/core/mocks/domains/upgrade.handlers.ts
@@ -1,6 +1,6 @@
 import { rest } from 'msw';
 
-import umbracoPath from '../../helpers/umbraco-path';
+import umbracoPath from '../../utils/umbraco-path';
 import type { PostInstallRequest, UmbracoUpgrader } from '@umbraco-cms/models';
 
 export const handlers = [
