+++ src/installer/consent/installer-content.test.ts
@@ -1,6 +1,6 @@
 import { expect, fixture, html } from '@open-wc/testing';
 
-import { defaultA11yConfig } from '../../core/helpers/chai';
+import { defaultA11yConfig } from '../../core/test-utils/chai';
 import { UmbInstallerConsentElement } from './installer-consent.element';
 
 // TODO: Write tests
