+++ src/installer/database/installer-database.test.ts
@@ -1,6 +1,6 @@
 import { expect, fixture, html } from '@open-wc/testing';
 
-import { defaultA11yConfig } from '../../core/helpers/chai';
+import { defaultA11yConfig } from '../../core/test-utils/chai';
 import { UmbInstallerDatabaseElement } from './installer-database.element';
 
 // TODO: Write tests
