+++ src/installer/installer.test.ts
@@ -1,6 +1,6 @@
 import { expect, fixture, html } from '@open-wc/testing';
 
-import { defaultA11yConfig } from '../core/helpers/chai';
+import { defaultA11yConfig } from '../core/test-utils/chai';
 import { UmbInstallerElement } from './installer.element';
 
 // TODO: Write tests
